Dev (#5)
增加非短信消息记录,增加redis缓存模板和账号

+++ app/msgpusher/internal/data/model/message_template.go
@@ -13,7 +13,7 @@ type MessageTemplate struct {
 	MsgType             int    `gorm:"column:msg_type" json:"msg_type"`                         // 10.通知类消息 20.营销类消息 30.验证码类消息
 	ShieldType          int    `gorm:"column:shield_type" json:"shield_type"`                   // 10.夜间不屏蔽 20.夜间屏蔽 30.夜间屏蔽(次日早上9点发送)
 	MsgContent          string `gorm:"column:msg_content" json:"msg_content"`                   // 消息内容 占位符用{$var}表示
-	SendAccount         int    `gorm:"column:send_account" json:"send_account"`                 // 发送账号 一个渠道下可存在多个账号
+	SendAccount         int64  `gorm:"column:send_account" json:"send_account"`                 // 发送账号 一个渠道下可存在多个账号
 	Creator             string `gorm:"column:creator" json:"creator"`                           // 创建者
 	Updator             string `gorm:"column:updator" json:"updator"`                           // 更新者
 	Auditor             string `gorm:"column:auditor" json:"auditor"`                           // 审核人
